Catégorie RI, juge 3, nommage et variable news Ajout catégorie RI et Juge 3 dans la config et les commandes Changement du nom des tickets pour futures fonctionnalités Variable news par défaut en cas de reset
