+++ config-vars.json
@@ -1,5 +1,5 @@
 {
-  "news": "Les inscriptions au concours sont ouvertes jusqu'au 13 janvier à 20h (UTC+1). \nMerci de lire le règlement attentivement avant de vous inscrire. \n\n Vous pouvez également [rejoindre le jury](https://discord.gg/ShGDXwt5BN) si vous êtes très motivé (tous les tests doivent être terminés le 20 janvier).  \n\n Si votre manuscrit est inscrit, les résultats des présélections seront donnés par mail le 27 janvier dans la journée. Pensez bien à vérifier vos spams !",
+  "news": "Vous pouvez suivre nos actualités sur notre site internet et nos réseaux sociaux, ainsi que grâce à notre newsletter.",
   "quilitquoi": "",
   "infoanim": "",
   "infonotif": "",
