+++ config.json
@@ -77,6 +77,13 @@
             "huberror": "1113909872069709916"
         },
         {
+            "id" : "ri",
+            "name" : "Romance Imaginaire",
+            "ticketparent" : "1318165115685572668",
+            "hubparent" : "1318168301695926343",
+            "huberror": "1318168436068581398"
+        },
+        {
             "id" : "hi",
             "name" : "Historique",
             "ticketparent" : "974645447513763840",
@@ -104,6 +111,12 @@
             "color" : "E99F86",
             "icon" : "https://les-murmures-litteraires.fr/wp-content/uploads/2024/01/mlbot_2.png"
         },
+        {
+            "id" : "3",
+            "name" : "Juge 3 de la première étape de lecture",
+            "color" : "FF5757",
+            "icon" : "https://les-murmures-litteraires.fr/wp-content/uploads/2024/12/mlbot_3.png"
+        },        
         {
             "id" : "a",
             "name" : "Juge A de la seconde étape de lecture",
